Move common functionality into utils module and update affected code.
